Fastq2ReadmapGmaxV1
This script takes in your forward reads, reverse reads, 6bp adapter, and output directory and returns all intermittent files with a sorted bam being the final output. This program is intended to be used on the UMN MSI network where all software is loaded as modules. The only item that would need to be changed is line 166 where the hard-coded path to your Reference genome would be.